Move shared volume validation features from elasticsearch to common package
